Cleanup of models
Changes:
Removed _id fields for plot and bean schemas.
Changed _id field for player to be unique and represent name of player.
Added 'virtual' field name, along with setter and getter for player.

Note:
I considered doing the same to game as I did to plot but upon reflection decided that we might want to use the timestamp in the default _id later.